--- core/__init__.py
@@ -1,5 +1,0 @@
-from .activity import *
-from .database import *
-from .github import *
-from .schema import *
-from .i18n import *
